// backend/src/common/enums/qc.enum.ts
export enum QcSampleStatus {
  PENDING = 'PENDING',
  IN_PROGRESS = 'IN_PROGRESS',
  COMPLETED = 'COMPLETED',
}

export enum QcTestName {
  ASSAY = 'ASSAY',
  DISSOLUTION = 'DISSOLUTION',
  PH = 'PH',
  RELATED_SUBSTANCES = 'RELATED_SUBSTANCES',
  IMPURITY = 'IMPURITY',
  IDENTIFICATION = 'IDENTIFICATION',
  MICROBIOLOGY = 'MICROBIOLOGY',
  STABILITY = 'STABILITY',
  APPEARANCE = 'APPEARANCE',
  WEIGHT_VARIATION = 'WEIGHT_VARIATION',
}

export enum QcTestResultStatus {
  PENDING = 'PENDING',
  PASS = 'PASS',
  FAIL = 'FAIL',
}